requirements.txt
tb-mqtt-client
pyserial
mmh3